Attempt to generalise the way we count and aggregate the metrics
This logic is still split out in many places. e.g. the statistics module
in the simulation still needs to know what metric is what, in order to
report the value. I'm not sure whether we will ever get away from this,
without providing the request into the metrics.

We ideally never want to provide the request into the metrics since some
metrics (e.g. utilisation) is not based on the request, but the
available time for the server.