Changes how we were scheduling the simulations to be more safe
I managed to kill a stream doing this, since we used to just schedule as
many threads as possible and leave it to the kernel to worry about the
handling. By including rayon, we can delegate the running of sims to a
thread pool which will handle this a lot more safely than what we were
doing previously (maybe at a cost though).