Remove smart pointers for Perf reasons
Which perf reasons? Well in theory non of the data that is ever provided
to a simulation should ever change after being provided to the sim.
Since RC's are not a zero cost abstraction, and the data is never
changed, then we don't "need" them. In terms of memory, the sims are
scaled to threads, so there shouldn't be any uncontrolled memory growth
(so long as the sims themselves remain light).

There will likely be regrets on this later. 👋 Future Ben.