Create Config namespace for Client/Server Profiles
This also means attribute is more shared than previously, and brings
this up into a more global namespace.

The idea behind this is to use ClientProfile, to generate Clients, and
create ServerProfile to generate Servers. That would make the
simulations a lot more simple in terms of generation of data.