Fairly large refactor of Metrics to support new RequestData
This now diverts the responsibility to the Metric itself to calculate.
Additionally we now have proper target condition checking.

There is a lot of match magic happening in this commit which would be
much nicer to catch at compile with some stricter type matching, but
then we start getting even more polymorphic, and that hurts my brain